"Add metadata and update LoginPage component for admin and cashier pages"
This diff includes changes to the admin and cashier pages in the 'app/users' directory. Both pages have a new Metadata object added to their respective files, setting the title to 'Admin' and 'Cashier' respectively. The LoginPage component has also been updated in both files to return a heading with the corresponding title.